piezo-dashboard/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
websockets==10.4
pyserial==3.5
python-multipart==0.0.5
pydantic==1.8.2
